+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 I'm following [this roadmap](https://dystroy.org/blog/bacon-everything-roadmap/). If you're interested in having bacon usable for other languages than Rust, come have a chat.
 
+### Next
+- allow specify scroll-pages action with a floating point number - Fix #264
+
 <a name="v3.3.0"></a>
 ### v3.3.0 - 2024/11/16
 - bacon can now be launched without Cargo.toml file